feat: bevy-less support
