+++ Cargo.toml
@@ -25,6 +25,13 @@ exclude = [
 ]
 
 
+[features]
+default = ["bevy"]
+
+bevy = ["dep:bevy"]
+
+
+
 [dependencies]
 clap = { version = "4.5", features = ["derive"] }
 serde = { version = "1.0", features = ["derive"] }
@@ -34,6 +41,7 @@ serde_json = "1.0"
 git = "https://github.com/bevyengine/bevy.git"
 rev = "b79b8133c881942a42bb359f998bd8792c48c1bc"
 default-features = false
+optional = true
 
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
